+++ flutter_readium_platform_interface/lib/method_channel_flutter_readium.dart
@@ -220,8 +220,7 @@ class MethodChannelFlutterReadium extends FlutterReadiumPlatform {
   @override
   // TODO(native): add iOS/Android native handler for setComicAutoPan; until
   // then calling this on non-web will throw MissingPluginException.
-  Future<void> setComicAutoPan(bool enabled) =>
-      methodChannel.invokeMethod('setComicAutoPan', enabled);
+  Future<void> setComicAutoPan(bool enabled) => methodChannel.invokeMethod('setComicAutoPan', enabled);
 
   @override
   Future<void> audioEnable({AudioPreferences? prefs, Locator? fromLocator}) =>